requirements.txt
fastapi
uvicorn
pillow
python-multipart
geopy
transformers
torch
requests
flask
pytest
httpx
GPSPhoto
reverse_geocoder
pycountry
exifread
piexif
pandas
countryinfo
moviepy==1.0.3
torchvision
rapidfuzz